+++ gameAboutSomething/Player.h
@@ -3,11 +3,11 @@ class Player {
 public:
 	Player();
 
-	Player(int coordinates_x, int coordinates_y, int health, int damage);
+	Player(int coordinates_x, int coordinates_y, double health, int damage);
 
 	int coordinates_x;
 	int coordinates_y;
-	int health;
+	double health;
 	int damage;
 };
